+++ src/Cart/Changelog.md
@@ -7,6 +7,7 @@
 
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support
+- Added the `getCart()` method to the `CartItem` interface
 
 ## 5.x Series